refactor: module splits, remove unused, feature flag for dependency
