+++ pallets/evm-compat/rpc/runtime-api/src/lib.rs
@@ -1,3 +1,7 @@
+//! evm-compact-runtime-api
+//!
+//! This module contains various helper function for the client to work properly
+
 #![cfg_attr(not(feature = "std"), no_std)]
 
 use codec::Codec;
@@ -15,28 +19,37 @@ sp_api::decl_runtime_apis! {
 		AccountId: Codec,
 		Balance: Codec,
 	{
+
+		/// find the mapped AccoundId
 		fn source_to_mapped_address(source: H160) -> AccountId;
 
+		/// check whether this h160 has a backing proxy behind it
 		fn source_is_backed_by(source: H160) -> Option<AccountId>;
 
 		fn check_contract_is_evm_compat(contract_addr: AccountId) -> Option<H160>;
 
+		/// get chain_id
 		fn chain_id() -> u64;
 
+		/// balances of the h160 address, only returns accounts not contracts
 		fn balances(address: H160) -> U256;
 
 
 		fn block_hash(number: u32) -> H256;
 
+		/// read contract storage of a contract
 		fn storage_at(address: H160, index: U256,) -> H256;
 
+		/// nonce of the address
 		fn account_nonce(addrss: H160) -> U256;
 
+		/// try-run a transaction, used to get the estimated cost or return value
 		fn call(from: Option<H160>, target: Option<H160>, value: Balance, input: Vec<u8>, gas_limit: u64) ->  Result<(Balance, ExecReturnValue), DispatchError>;
 
+		/// get the block author, returns the first 20 bytes as h160 identifier
 		fn author(digest: Vec<ConesensusDigest>) -> Option<H160>;
 
-
+		/// return only extrinsics that contains valid eth-transaction
 		fn extrinsic_filter(
 			xts: Vec<<Block as BlockT>::Extrinsic>,
 		) -> Vec<TransactionV2>;
