+++ runtime/src/impl_pallet_contracts/chain_extensions.rs
@@ -3,7 +3,6 @@ use crate::Currencies;
 use codec::Encode;
 use frame_support::{
 	log::error,
-	sp_runtime,
 	sp_runtime::{traits::AccountIdConversion, DispatchError},
 };
 use frame_system::RawOrigin;
